requirements.txt
streamlit
requests
wordcloud
matplotlib
python-dotenv
